% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convertFeaturesToSortedMatrix}
\alias{convertFeaturesToSortedMatrix}
\title{Converts a set of features, as a FeatureMap, to a matrix of features.}
\usage{
convertFeaturesToSortedMatrix(features)
}
\arguments{
\item{features}{The set of features to convert, as a FeatureMap.}
}
\value{
The corresponding sorted matrix for the input features.
}
\description{
Each row of the matrix represents a feature, and contains its retention
time (RT), mass-to-charge (m/z), and signal intensity in that order. The
resulting matrix will be sorted by descending signal intensity.
}
\examples{
\dontrun{
read feature map
convert feature map
}
}
