% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{findFirstFeature}
\alias{findFirstFeature}
\title{Finds the first feature in a matrix that has its element at a given key
equal to a given target. If no feature matches the requirement, counts
how many features have the same key less than the target (its rank).}
\usage{
findFirstFeature(sortedFeatureMatrix, key = RT_IDX, target = 0)
}
\arguments{
\item{sortedFeatureMatrix}{The sorted feature matrix to search.}

\item{key}{The column in the matrix to search in.}

\item{target}{The target value to search for, as a number.}
}
\value{
The index of the first feature that has its element at the given key
equal to the given target, if it exists, and the rank of the matrix
otherwise.
}
\description{
This function uses binary search, so the feature matrix must be sorted
ascending in the key column.
}
\examples{
\dontrun{
m <- matrix(1:10, nrow = 5, ncol = 2)
m[2, 1] <- 3
m[4, 1] <- 3
rank <- findFirstFeature(m, key = 1, target = 3)
rank  # Should be 2
}
}
