% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{findFirstFeatureByMZ}
\alias{findFirstFeatureByMZ}
\title{Finds the first feature in a feature matrix that has its mass-to-charge
equal to a given target.}
\usage{
findFirstFeatureByMZ(featureMatrix, target)
}
\arguments{
\item{featureMatrix}{The feature matrix to search.}

\item{target}{The target m/z value, as a number.}
}
\value{
The first feature with the given m/z, if it exists; NULL otherwise.
}
\description{
This function searches the feature matrix by sorting it first, and then
applying binary search. If no feature with the specified m/z is found, then
NULL is returned.
}
\examples{
\dontrun{
features <- loadFeatureFile("inst/extdata/featureSetA.featureXML")
featureMatrix <- convertFeaturesToSortedMatrix(features)
rank <- findFirstFeatureByMZ(featureMatrix, 300)
rank  # TODO: what does this return?
}
}
