% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{findFirstFeatureByRT}
\alias{findFirstFeatureByRT}
\title{Finds the first feature in a feature matrix that has its retention time
equal to a given target.}
\usage{
findFirstFeatureByRT(featureMatrix, target)
}
\arguments{
\item{featureMatrix}{The feature matrix to search.}

\item{target}{The target retention time, as a number.}
}
\value{
The first feature with the given RT, if it exists; NULL otherwise.
}
\description{
This function searches the feature matrix by sorting it first, and then
applying binary search. If no feature with the specified RT is found, then
NULL is returned.
}
\examples{
\dontrun{
features <- loadFeatureFile("inst/extdata/featureSetA.featureXML")
featureMatrix <- convertFeaturesToSortedMatrix(features)
rank <- findFirstFeatureByRT(featureMatrix, 810)
rank  # TODO: what does this return?
}
}
