% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{generateFeatureDataFrame}
\alias{generateFeatureDataFrame}
\title{Data frame generator.}
\usage{
generateFeatureDataFrame(featureFilePath)
}
\arguments{
\item{featureFilePath}{The location of the featureXML file, as a string.}
}
\value{
The corresponding data frame for the feature set.
}
\description{
Generates a data frame from a feature set. Every feature is represented as a
row, containing its retention time, mass-to-charge, and signal intensity, in
that order.
}
\examples{
\dontrun{
featureFrame <- generateFeatureDataFrame(
"inst/extdata/featureSetA.featureXML")
}

}
