% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{generateMSDataFrame}
\alias{generateMSDataFrame}
\title{Mass spectrometry data frame generator.}
\usage{
generateMSDataFrame(rawDataFilePath)
}
\arguments{
\item{rawDataFilePath}{The location of the mzML file, as a string.}
}
\value{
The corresponding data frame for the dataset.
}
\description{
Generates a data frame from a raw mass spectrometry dataset. Every data
point/peak is represented as a row containing its retention time, mass-to-
charge, and signal intensity.
}
\examples{
\dontrun{
msFrame <- generateMSDataFrame(
"inst/extdata/20190122_HeLa_QC_Slot1-47_1_3228_800-810.mzML")
}

}
