% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparator.R
\name{getFeatureByIdx}
\alias{getFeatureByIdx}
\title{Specific feature retrieval.}
\usage{
getFeatureByIdx(featureFilePath, idx)
}
\arguments{
\item{featureFilePath}{The location of the featureXML file, as a string.}

\item{idx}{The index of the required feature in the set, as an integer.}
}
\value{
The specified feature's RT, m/z, and intensity, as a string.
}
\description{
Gets information (in particular, retention time, mass-to-charge, and signal
intensity) for the feature at the given index in the given feature file. If
the feature does not exist, an error is thrown.
}
\examples{
\dontrun{
getFeatureByIdx("inst/extdata/featureSetA.featureXML", 250)
}

}
