% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{identicalFeatures}
\alias{identicalFeatures}
\title{Compares two features to determine if they are identical.}
\usage{
identicalFeatures(feature1, feature2, rtThreshold = 5, mzThreshold = 0.01)
}
\arguments{
\item{feature1}{The first feature to compare}

\item{feature2}{The second feature to compare}

\item{rtThreshold}{The retention time threshold to use}

\item{mzThreshold}{The mass-to-charge threshold to use}
}
\value{
TRUE if the features are identical; FALSE otherwise.
}
\description{
Features can be considered to be identical if their retention times and
mass-to-charges are within some thresholds of each other (which can be
provided by the user).
}
\examples{
\dontrun{
featureA <- ropenms$Feature()
featureA$setRT(5)
featureA$setMZ(300)
featureB <- ropenms$Feature()
featureB$setRT(6)
featureB$setMZ(300.005)
identicalFeatures(featureA, featureB)  # TRUE
}
}
