% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotFeatureSetOnRawData}
\alias{plotFeatureSetOnRawData}
\title{Layered feature set plotter.}
\usage{
plotFeatureSetOnRawData(rawDataFilePath, featureFilePath)
}
\arguments{
\item{rawDataFilePath}{The location of the mzML file, as a string.}

\item{featureFilePath}{The location of the featureXML file, as a string.}
}
\value{
The overlayed feature/raw data scatter plot.
}
\description{
Creates a scatter plot, where the given feature set is layered on top of its
raw mass spectrometry dataset. The data is distinguished from the features
by colour.
}
\details{
The x-axis represents retention time (in sec) and the y-axis represents
mass-to-charge (in Th).

This function may take a while to run.
}
\examples{
\dontrun{
plotFeatureSetOnRawData(
"inst/extdata/20190122_HeLa_QC_Slot1-47_1_3228_800-810.mzML",
"inst/extdata/featureSetA.featureXML")
}

}
