% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotRawData}
\alias{plotRawData}
\title{Raw data plotter.}
\usage{
plotRawData(rawDataFilePath)
}
\arguments{
\item{rawDataFilePath}{The location of the mzML file, as a string.}
}
\value{
Nothing. The plot is outputted to the screen.
}
\description{
Plots a raw mass spectrometry dataset onto a scatter plot.
}
\details{
The x-axis represents retention time (in sec), the y-axis represents mass-
to-charge (in Th), and the z-axis (with darker colours for greater values)
represents signal intensity (unitless).

This function may take a while to run.
}
\examples{
\dontrun{
plotRawData("inst/extdata/20190122_HeLa_QC_Slot1-47_1_3228_800-810.mzML")
}

}
