% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotSingleFeatureSet}
\alias{plotSingleFeatureSet}
\title{Plots a feature set onto a blank scatter plot with colour.}
\usage{
plotSingleFeatureSet(featureFilePath)
}
\arguments{
\item{featureFilePath}{The location of the featureXML file, as a string.}
}
\description{
The x-axis represents retention time, the y-axis represents mass-to-charge,
and the z-axis (the colour) represents intensity.
}
\details{
This function may take a while to run.
}
\examples{
\dontrun{
plotSingleFeatureSet("inst/extdata/featureSetA.featureXML")
}

}
