% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotTwoFeatureSets}
\alias{plotTwoFeatureSets}
\title{Plots two feature sets (overlapping) onto a blank scatter plot,
distinguished by colour.}
\usage{
plotTwoFeatureSets(featureFilePath1, featureFilePath2)
}
\arguments{
\item{featureFilePath1}{The location of the first featureXML file, as a
string.}

\item{featureFilePath2}{The location of the second featureXML file, as a
string.}
}
\description{
The x-axis represents retention time and the y-axis represents mass-to-
charge.
}
\details{
This function may take a while to run.
}
\examples{
\dontrun{
plotTwoFeatureSets("inst/extdata/featureSetA.featureXML",
"inst/extdata/featureSetB.featureXML")
}

}
