% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMsFeatureCmp.R
\name{runMsFeatureCmp}
\alias{runMsFeatureCmp}
\title{Shiny app launcher}
\usage{
runMsFeatureCmp()
}
\value{
Nothing. The shiny app is launched.
}
\description{
Launches the shiny app for the package (msFeatureCmp). The shiny code is
located in \code{./inst/shiny-scripts}.
}
\examples{
\dontrun{
runMsFeatureCmp()
}

}
