% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{similarFeatures}
\alias{similarFeatures}
\title{Compares two features to determine if they are similar.}
\usage{
similarFeatures(feature1, feature2, rtThreshold = 5, mzThreshold = 0.01)
}
\arguments{
\item{feature1}{The first Feature to compare.}

\item{feature2}{The second Feature to compare.}

\item{rtThreshold}{The RT threshold to use, as a number.}

\item{mzThreshold}{The m/z threshold to use, as a number.}
}
\value{
TRUE if the two features are similar; FALSE otherwise.
}
\description{
We consider features to be similar if their retention times and mass-to-
charge values are within some fixed thresholds of each other (which can be
provided by the user).
}
\examples{
\dontrun{
featureA <- ropenms$Feature()
featureA$setRT(5)
featureA$setMZ(300)
featureB <- ropenms$Feature()
featureB$setRT(6)
featureB$setMZ(300.005)
similarFeatures(featureA, featureB)  # Returns TRUE
}
}
