% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sortMatrixByColumn}
\alias{sortMatrixByColumn}
\title{Sorts a matrix by a given column number.}
\usage{
sortMatrixByColumn(matrix, column = 1, descending = FALSE)
}
\arguments{
\item{matrix}{The matrix to sort.}

\item{column}{The column number to use as key.}

\item{descending}{If TRUE, sorts the matrix by descending values.}
}
\value{
The sorted matrix.
}
\description{
Sorts a matrix by a given column number.
}
\examples{
\dontrun{
m <- matrix(1:9, nrow = 3, ncol = 3)
m[1, 1] <- 2
m[2, 1] <- 1
m[2, 2] <- 6
m[3, 2] <- 5
sortedMatrix <- sortMatrixByColumn(m, descending = TRUE)
sortedMatrix
}
}
