% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{withinThreshold}
\alias{withinThreshold}
\title{Threshold checker.}
\usage{
withinThreshold(value1, value2, threshold)
}
\arguments{
\item{value1}{The first value, as a number.}

\item{value2}{The second value, as a number.}

\item{threshold}{The maximum distance between the values, as a number.}
}
\value{
TRUE if the values are within the distance; FALSE otherwise.
}
\description{
Checks if two values are within a certain one-dimensional distance of each
other.
}
\examples{
\dontrun{
x1 <- 5.1
x2 <- 5.05
maxDist <- 0.01
withinThreshold(x1, x2, maxDist)  # Returns FALSE
}
}
