% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparator.R
\name{compareFeatures}
\alias{compareFeatures}
\title{Feature set comparator.}
\usage{
compareFeatures(rawDataFilePath, featureFilePath1, featureFilePath2)
}
\arguments{
\item{rawDataFilePath}{The location of the mzML raw data file, as a string.}

\item{featureFilePath1}{The location of the first featureXML file, as a
string.}

\item{featureFilePath2}{The location of the second featureXML file, as a
string.}
}
\value{
Nothing. A basic analysis is printed to the screen instead.
}
\description{
Compares two sets of mass spectrometry features corresponding to a single MS
run. This can be used to compare the results of different feature finding
algorithms that have been run on the same raw dataset.
}
\details{
A short summary, including the number of features that were found to be
common to both sets of features, as well as statistical values such as
recall, precision, and F_1 score, will be printed to the terminal.

This function may take a while to run.
}
\examples{
\dontrun{
compareFeatures("inst/extdata/20190122_HeLa_QC_Slot1-47_1_3228_800-810.mzML",
                "inst/extdata/featureSetA.featureXML",
                "inst/extdata/featureSetB.featureXML")
}

}
