% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{compareFeatures}
\alias{compareFeatures}
\title{Compares two sets of mass spectrometry features on a raw mass spectrometry
data set.}
\usage{
compareFeatures(featureFilePath1, featureFilePath2, rawDataFilePath)
}
\arguments{
\item{featureFilePath1}{The location of the first featureXML file.}

\item{featureFilePath2}{The location of the second featureXML file.}

\item{rawDataFilePath}{The location of the mzML file.}
}
\description{
Compares two sets of mass spectrometry features on a raw mass spectrometry
data set.
}
\examples{
\dontrun{
compareFeatures("inst/extdata/featureSetA.featureXML",
                "inst/extdata/featureSetB.featureXML",
                "inst/extdata/20190122_HeLa_QC_Slot1-47_1_3228_800-860.mzML")
}

}
