% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparator.R
\name{compareFeatures}
\alias{compareFeatures}
\title{Compares two sets of mass spectrometry features corresponding to a single MS
run. For example, to compare the results of different feature finding
algorithms run on the same raw dataset.}
\usage{
compareFeatures(featureFilePath1, featureFilePath2, rawDataFilePath)
}
\arguments{
\item{featureFilePath1}{The location of the first featureXML file, as a string.}

\item{featureFilePath2}{The location of the second featureXML file, as a string.}

\item{rawDataFilePath}{The location of the mzML file, as a string.}
}
\description{
TODO: it should print a comparison, including statistical scores. What else?
}
\examples{
\dontrun{
compareFeatures("inst/extdata/featureSetA.featureXML",
                "inst/extdata/featureSetB.featureXML",
                "inst/extdata/20190122_HeLa_QC_Slot1-47_1_3228_800-860.mzML")
}

}
