% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convertFeaturesToSortedMatrix}
\alias{convertFeaturesToSortedMatrix}
\title{Converts a loaded set of features into a matrix of features.}
\usage{
convertFeaturesToSortedMatrix(featureSet)
}
\arguments{
\item{featureSet}{The set of features to convert, as a FeatureMap.}
}
\value{
The corresponding sorted matrix for the feature set.
}
\description{
Each row of the matrix represents a feature, and contains its retention
time, mass-to-charge, and signal intensity values, in that order. The matrix
will be also sorted by ascending RT.
}
\examples{
\dontrun{
featureSet <- loadFeatureFile("inst/extdata/featureSetA.featureXML")
featureMatrix <- convertFeaturesToSortedMatrix(featureSet)
featureMatrix[1, ]
# Returns [800.002794376452, 614.596298389786, 10900.7001953125]
}
}
