% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{findFirstFeatureByRT}
\alias{findFirstFeatureByRT}
\title{Finds the first feature in a feature matrix that has its retention time equal
to a given target. If no such feature exists, counts how many features have
retention time less than the target (the matrix rank).}
\usage{
findFirstFeatureByRT(featureMatrix, target)
}
\arguments{
\item{featureMatrix}{The feature matrix to search.}

\item{target}{The target retention time, as a number.}
}
\value{
The index of the first feature that has its retention time equal to
the given target, if it exists, and the rank of the matrix otherwise.
}
\description{
This function searches the feature matrix by sorting it first, and then
applying binary search.
}
\examples{
\dontrun{
features <- loadFeatureFile("inst/extdata/featureSetA.featureXML")
featureMatrix <- convertFeaturesToSortedMatrix(features)
rank <- findFirstFeatureByRT(featureMatrix, 810)
rank  # TODO: what does this return?
}
}
