% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparator.R
\name{getFeatureByIdx}
\alias{getFeatureByIdx}
\title{Retrieves information about the feature at the given index in the given
featureXML file.}
\usage{
getFeatureByIdx(featureFilePath, idx)
}
\arguments{
\item{featureFilePath}{The location of the featureXML file, as a string.}

\item{idx}{The index of the required feature in the set, as an integer.}
}
\description{
In particular, this function prints the feature's retention time, mass-to-
charge, and signal intensity, if it exists, and throws an error otherwise.
}
\examples{
\dontrun{
getFeatureByIdx("inst/extdata/featureSetA.featureXML", 250)
}

}
