% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{loadFeatureFile}
\alias{loadFeatureFile}
\title{Loads a mass spectrometry feature set into memory. The feature file must be
featureXML format.}
\usage{
loadFeatureFile(filePath)
}
\arguments{
\item{filePath}{The location of the featureXML file.}
}
\value{
The in-memory representation of the feature set, as a FeatureMap.
}
\description{
Loads a mass spectrometry feature set into memory. The feature file must be
featureXML format.
}
\examples{
\dontrun{
featureSet <- loadFeatureFile("inst/extdata/featureSetA.featureXML")
}
}
