% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{loadMSFile}
\alias{loadMSFile}
\title{Loads a raw mass spectrometry run into memory. The MS file must be mzML
format.}
\usage{
loadMSFile(filePath)
}
\arguments{
\item{filePath}{The location of the mzML file.}
}
\value{
The in-memory representation of the MS run, as an MSExperiment.
}
\description{
Loads a raw mass spectrometry run into memory. The MS file must be mzML
format.
}
\examples{
\dontrun{
experiment <- loadMSFile("inst/extdata/20190122_HeLa_QC_Slot1-47_1_3228_800-860.mzML")
}
}
