% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotFeatureSetOnRawData}
\alias{plotFeatureSetOnRawData}
\title{Plots a feature set layered on top of its raw mass spectrometry dataset,
distinguished by colour.}
\usage{
plotFeatureSetOnRawData(rawDataFilePath, featureFilePath)
}
\arguments{
\item{rawDataFilePath}{The location of the mzML file, as a string.}

\item{featureFilePath}{The location of the featureXML file, as a string.}
}
\description{
The x-axis represents retention time and the y-axis represents mass-to-
charge.
}
\details{
This function may take a while to run.
}
\examples{
\dontrun{
plotFeatureSetOnRawData(
"inst/extdata/20190122_HeLa_QC_Slot1-47_1_3228_800-810.mzML",
"inst/extdata/featureSetA.featureXML")
}

}
