% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotRawData}
\alias{plotRawData}
\title{Plots a raw mass spectrometry dataset onto a scatter plot with colour.}
\usage{
plotRawData(rawDataFilePath)
}
\arguments{
\item{rawDataFilePath}{The location of the mzML file, as a string.}
}
\description{
The x-axis represents retention time, the y-axis represents mass-to-charge,
and the z-axis (the colour) represents intensity.
}
\details{
This function may take a while to run.
}
\examples{
\dontrun{
plotRawData("inst/extdata/20190122_HeLa_QC_Slot1-47_1_3228_800-810.mzML")
}

}
