% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotSingleFeatureSet}
\alias{plotSingleFeatureSet}
\title{Single feature set plotter.}
\usage{
plotSingleFeatureSet(featureFilePath)
}
\arguments{
\item{featureFilePath}{The location of the featureXML file, as a string.}
}
\value{
The feature set scatter plot.
}
\description{
Plots a single feature set onto a blank scatter plot.
}
\details{
The x-axis represents retention time (in sec), the y-axis represents mass-
to-charge (in Th), and the z-axis (with darker colours for greater values)
represents signal intensity (unitless).

This function may take a while to run.
}
\examples{
\dontrun{
plotSingleFeatureSet("inst/extdata/featureSetA.featureXML")
}

}
