% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotTwoFeatureSets}
\alias{plotTwoFeatureSets}
\title{Overlapping feature set plotter.}
\usage{
plotTwoFeatureSets(featureFilePath1, featureFilePath2)
}
\arguments{
\item{featureFilePath1}{The location of the first featureXML file, as a
string.}

\item{featureFilePath2}{The location of the second featureXML file, as a
string.}
}
\value{
Nothing. The plot is outputted to the screen.
}
\description{
Plots two feature sets, at the same time, onto a blank scatter plot. The two
feature sets are distinguished by colour.
}
\details{
The x-axis represents retention time (in sec) and the y-axis represents
mass-to-charge (in Th).

This function may take a while to run.
}
\examples{
\dontrun{
plotTwoFeatureSets("inst/extdata/featureSetA.featureXML",
"inst/extdata/featureSetB.featureXML")
}

}
