% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{similarFeatures}
\alias{similarFeatures}
\title{Feature similarity checker.}
\usage{
similarFeatures(
  feature1,
  feature2,
  rtThreshold = RT_THRESHOLD,
  mzThreshold = MZ_THRESHOLD
)
}
\arguments{
\item{feature1}{The first feature to compare.}

\item{feature2}{The second feature to compare.}

\item{rtThreshold}{The RT threshold to use, as a number.}

\item{mzThreshold}{The m/z threshold to use, as a number.}
}
\value{
TRUE if the two features are similar; FALSE otherwise.
}
\description{
Compares two features to determine if they are similar.
}
\details{
We consider features to be similar if their retention times and mass-to-
charge values are within some fixed threshold of each other, which can be
determined and provided by the user.

The provided features can be ropenms Feature objects or vectors (e.g. from
a feature matrix).
}
\examples{
\dontrun{
featureA <- ropenms$Feature()
featureA$setRT(5)
featureA$setMZ(300)
featureB <- ropenms$Feature()
featureB$setRT(6)
featureB$setMZ(300.005)
similarFeatures(featureA, featureB)  # Returns TRUE
}

}
