% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{similarFeatures}
\alias{similarFeatures}
\title{Compares two features to determine if they are similar.}
\usage{
similarFeatures(feature1, feature2, rtThreshold = 5, mzThreshold = 0.01)
}
\arguments{
\item{feature1}{The first feature to compare, as a Feature.}

\item{feature2}{The second feature to compare, as a Feature.}

\item{rtThreshold}{The retention time threshold to use, as a float.}

\item{mzThreshold}{The mass-to-charge threshold to use, as a float.}
}
\value{
TRUE if the two features are similar; FALSE otherwise.
}
\description{
We consider features to be similar if their retention times and mass-to
charge values are within some fixed thresholds of each other (which can be
provided by the user).
}
\details{
The default thresholds have been experimentally determined to provide the
"best" results.
}
\examples{
\dontrun{
featureA <- ropenms$Feature()
featureA$setRT(5)
featureA$setMZ(300)
featureB <- ropenms$Feature()
featureB$setRT(6)
featureB$setMZ(300.005)
similarFeatures(featureA, featureB)  # Returns TRUE
}
}
